Remove ExceptT fro SKC
Modify `SKC', remove `ExceptT' to simplify the code. Remove variant
functions in Builder and Lexer which were lifting the `Either' result to
`ExceptT'.

Add `SkException' newtype, made as an instance of Exception typeclass.

Modify definition of `failS' to throw `SkException'. Modify codes in Run
in Plugin to work with modified codes.