Do not add macro to compiler context with define-macro
Modify `define-macro` special form to not add resulting macro to current
compilation context. To add the resulting macro to current compiler
context, wrap the `define-macro` form with `eval-when-compile`. Add
`envAddInDefineMacro` field to SkEnv to configure this behavior and add
the resulting macro with `define-macro` to current context.

Modify tests for macro to meet the modification in `define-macro`.