Update documentation for building sample packages
Relates to #1. Modify "Building Package" section of the documentation
to reflect recent changes. Modify the codes literally included in the
documentation.

Add tips for how to build a package with the latest finkel package
from source codes in the git repository.

Add a warning section to refer that the "stack new" command cannot
find finkel dependency package because the packages are not yet
uploaded to Stackage and Hackage.

Replace "finkel-core" package name with "finkel-lang", "finkel-core"
was the old name for "finkel-lang".