Refine REPL interaction for inputs
Modify REPL listener to avoid infinite loop when network connection
has been closed.

Modify "read-print-loop" to interrupt with "withInterrupt" when the
user hit C-c in the middle of entering a form, to support resetting
malformed input. Move the call to "withInterrupt" wrapping the entire
loop to asynchronous communication for long computation work.