Num and Fractional instance for Code
Add instance of "Num" and "Fractional" type classs to "Code". Evaluates to
"AInteger" when the arguments were "AInteger", "AFractional" when the arguments
were "AFractional". For binary operations, evaluates to "AFractional" if any of
the arguments were "AFractional". Any other values than "AInteger" or
"AFractional" will results to "nil".

Add simple tests for "Num" and "Fractional" methods.