Add in-package REPL command
Add "in-package" to REPL commands, to support setting the "thisInstalledUnitId"
of the current "DynFlags" to an "InstalledUnitId" specified by name. Add simple
test for the "in-package" command.

By setting the "thisInstalledUnitId", REPL can reload the modules from source
codes instead of compiled objects and libraries when working with codes in
cabalized packages.

Some minor tidy ups.