Use bytecode and object code for modules import in macro expansion
Modify macro expansion, detect whether to use object code or not with
"canExpandWithObj", reuse the "HscEnv" from the main object code compilation
work.  Modify ":require" and ":eval-when-compile" to use the bytecode
interpreter when importing modules when not using object code.

Modify "withExpanderSettings" to decide whether to use the dedicated "HscEnv"
for bytecode interpreter by looking the settings in current "DynFlags".

Modify "checkUpToDate", return "False" when current compilation target is object
code when the target "ModSummary" were not found.

Remove "NoMonomorphismRestriction" and set the "ghcLink" to "LinkInMemory" in
"setExpanderSettings", to make the bytecode interpreter to behave with default
"DynFlags" settings.

Modify tests in "EvalTest", add "Prelude" and "Language.Finkel" modules to the
interactive context, so that the bytecode interpreter used during macro
expansion can refer to the codes from the modules.