Update configuration for GitHub workflow
Replace the use of "::add-path::" with echoing to "$GITHUB_PATH", since the
messages from the GitHub runner were saying that the "::add-path::" command will
be disabled. Modify the command arguments to "ghcup", the output message from
the GitHub runner was showing that the command line arguments were written in
old style.