Get ghc top level directory at runtime by default
Modify "defaultFnkEnv", use "Nothing" as the value of "envLibDir".

Modify "runFnk" to get ghc top level directory at runtime if the "envLibDir"
field in "FnkEnv" was "Nothing". This change will enable running the "Fnk"
actions with other than the environment used for compiling the "finkel-kernel"
package, which should be more general than using the absolute path.

Add "fnkTestEnv" for tests in "finkel-kernel" package to use the ghc top level
directory used for compiling the package.