Do not expand list forms starting with `:quote'
Modify "Language.Finkel.Expand.expand'" to return the original form when the
form was a list starting with ":quote".

Add a test for `macroexpand-1' to expand a macro which expands to a form
containing `(:begin ...)'.

Add "expand-form-with-env" in test auxiliary function, to use given "FnkEnv"
from argument, to update test macros used during macro expansion.