Drop support for ghc < 8.10 in finkel-tool
Modify cond-expand used for supporting ghc < 8.10.

Modify lower version of dependency packages to bundled packages of ghc
8.10. Remove ghc-options used for compiling with ghc < 8.10 in library and test
stanzas.

Simplify Setup.hs, drop the custom user hooks used for supporting ghc < 8.10.

Remove '(:require Finkel.Core)' in the beginning of modules, this macro is used
to support compiling with 'fkc', which is used in ghc < 8.10.